:art: refactor: Move encoder to GameSolver.Collection.Encoder namespace
